chatbot project/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langgraph
chromadb
sentence-transformers
huggingface-hub
tavily-python
pydantic
python-multipart
python-dotenv

